move Discord Webhook to lib from classes
