+++ src/lib/DiscordWebhook/utils.ts
@@ -1,4 +1,4 @@
-import Bot from '../Bot';
+import Bot from '../../classes/Bot';
 import TradeOfferManager, { TradeOffer } from 'steam-tradeoffer-manager';
 
 export function getPartnerDetails(offer: TradeOffer, bot: Bot, callback: (err: any, details: any) => void): any {
